+++ src/grid.h
@@ -1,30 +1,31 @@
 #pragma once
 
+#include <cstdint>
 #include <optional>
 #include <vector>
 #include <string>
 #include <fstream>
 
 struct Tile {
-    unsigned int type;
+    uint16_t type;
 };
 
 struct Grid{
-    unsigned int size_x;
-    unsigned int size_y;
+    uint16_t size_x;
+    uint16_t size_y;
     //TODO: should probably be immutable, but reassignable
 
     std::vector<std::vector<Tile>> tiles;
 
     Grid(size_t width, size_t height);
 
-    void Place(unsigned int x, unsigned int y, unsigned int type);
+    void Place(uint16_t x, uint16_t y, uint16_t type);
 
-    Tile GetTile(unsigned int x, unsigned int y) const; //TODO: Maybe make this a reference idk
+    Tile GetTile(uint16_t x, uint16_t y) const; //TODO: Maybe make this a reference idk
 
     void SaveToFile(std::string filename);
 
     static std::optional<Grid> LoadFromFile(std::string filename);
 
-    static Grid NewDefault(unsigned int width, unsigned int height);
+    static Grid NewDefault(uint16_t width, uint16_t height);
 };
